# cmd/core/model/agent.go
package model

import "encoding/json"

type AgentID string

type Agent struct {
	ID   AgentID `json:"id"`
	Name string  `json:"name"`
}

type AgentSkill struct {
	Name    string          `json:"name"`
	Payload json.RawMessage `json:"payload" swaggertype:"object"`
}

type EventSource struct {
	Name string `json:"name"`
}