# docs/swagger.yaml
basePath: /
definitions:
  agentsdk.CameraConfig:
    properties:
      camera_id:
        type: string
      type:
        $ref: '#/definitions/agentsdk.CameraType'
      url:
        type: string
    type: object
  agentsdk.CameraType:
    enum:
    - mjpg
    - rtsp
    - webrtc
    type: string
    x-enum-varnames:
    - CameraTypeMJpg
    - CameraTypeRtsp
    - CameraTypeWebRtc
  agentsdk.RegisterResponse:
    properties:
      grpc_addr:
        type: string
      token:
        type: string
    type: object
  controller.agentView:
    properties:
      cameras:
        items:
          $ref: '#/definitions/agentsdk.CameraConfig'
        type: array
      id:
        type: string
      name:
        type: string
      online:
        type: boolean
      skills:
        items:
          $ref: '#/definitions/model.AgentSkill'
        type: array
    type: object
  controller.listAgentsResponse:
    properties:
      agents:
        items:
          $ref: '#/definitions/controller.agentView'
        type: array
    type: object
  controller.registerRequest:
    properties:
      cameras:
        items:
          $ref: '#/definitions/agentsdk.CameraConfig'
        type: array
      event_sources:
        items:
          $ref: '#/definitions/model.EventSource'
        type: array
      skills:
        items:
          $ref: '#/definitions/model.AgentSkill'
        type: array
    type: object
  model.AgentSkill:
    properties:
      name:
        type: string
      payload:
        type: object
    type: object
  model.EventSource:
    properties:
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: VantageOS backend — agentsdk registry and task management.
  title: VantageOS Core API
  version: "1.0"
paths:
  /agents:
    get:
      description: |-
        Returns all provisioned agents, each annotated with whether it
        currently has a live gRPC stream, plus its reported skills and cameras.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.listAgentsResponse'
      summary: List agents
      tags:
      - agents
  /agents/register:
    post:
      consumes:
      - application/json
      description: |-
        Called by a physical agentsdk on boot to register its identity and skills.
        The request must include a pre-shared device API key as a Bearer token.
      parameters:
      - description: Bearer <device-api-key>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Agent registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agentsdk.RegisterResponse'
        "400":
          description: bad request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Register an agentsdk
      tags:
      - agents
swagger: "2.0"